+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # 2.x branch
+## 2.4 branch
+### 2.4.0
+* fixed bug trying to email a nonexistent backup;
+* `VALID_COMPRESSIONS` and `VALID_EXTENSIONS` constants have been replaced by
+    `getValidCompressions()` and `getValidExtensions()` methods provided by the
+    `BackupTrait` class;
+* replaced `InternalErrorException` with `InvalidArgumentException` and 
+    `RuntimeException`. This allows compatibility with CakePHP 3.6 branch.
+
 ## 2.3 branch
 ### 2.3.0
 * full support for working under Windows;